fixed heredoc file cannot be found when control d by adding a flag to free_execution, fixed pipex unconditional jump issue by initialising the status in the recus function, fixed variable expansion issue with =s

+++ srcs/minishell/utils/free_mini.c
@@ -12,7 +12,7 @@
 
 #include "../../../includes/minishell.h"
 
-void	free_t_parse(t_parse *ptr)
+void	free_t_parse(t_parse *ptr, int type)
 {
 	t_parse	*current;
 	t_parse	*next;
@@ -25,7 +25,7 @@ void	free_t_parse(t_parse *ptr)
 		next = current->next;
 		if (current->heredoc != NULL)
 		{
-			if (access(current->heredoc, F_OK) == 0)
+			if (type != HDOC && access(current->heredoc, F_OK) == 0)
 				unlink(current->heredoc);
 			free(current->heredoc);
 		}
@@ -70,14 +70,18 @@ void	free_t_export(t_export *list)
 	}
 }
 
-void	free_execution(t_mini *mini)
+void	free_execution(t_mini *mini, int type)
 {
-	free_t_parse(mini->input);
+	free_t_parse(mini->input, type);
 	free_t_export(mini->list);
 	free_t_env(mini->env);
 	close(mini->term_in);
 	close(mini->term_out);
 	close(mini->std_err);
+	if (mini->in != -1)
+		close(mini->in);
+	if (mini->out != -1)
+		close(mini->out);
 	free(mini->user_input);
 	free(mini);
 }
@@ -101,6 +105,6 @@ void	ft_free_all(t_mini *mini, int state)
 	}
 	if (mini->exit == 0)
 		free(mini->user_input);
-	free_t_parse(mini->input);
+	free_t_parse(mini->input, 0);
 	mini->input = NULL;
 }
